Move URLWhitelistInput to common folder
