+++ graylog2-web-interface/src/components/common/URLWhitelistInput.jsx
