+++ graylog2-web-interface/src/components/event-notifications/event-notification-types/HttpNotificationForm.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import lodash from 'lodash';
-import URLWhitelistInput from 'components/configurations/URLWhitelistInput';
+import URLWhitelistInput from 'components/common/URLWhitelistInput';
 import FormsUtils from 'util/FormsUtils';
 
 class HttpNotificationForm extends React.Component {
